requirements.txt
amqp
celery
Flask
flower
gunicorn
humanize
itsdangerous
Jinja2
MarkupSafe
pytz
redis
six
Werkzeug
openai
python-dotenv
flask-sqlalchemy
psycopg2